requirements.txt
requests==2.31.0
PyYAML==6.0.1
python-dotenv==1.0.0
pytest==7.4.0
